Add jni load calls only for shared library buils
This should fix #50,

If users do not define their own functions, the generator will generate
a file that needs to be inlcuded.
For an own JNI_OnLoad/JNI_OnUnload function, the generator needs to be
informed.

If the support lib is used as a dynamic lib, the JNI functions are not
optional.